requirements.txt
PyQt6>=6.4.0
pycryptodome>=3.15.0
bcrypt>=4.0.0
